+++ src/plot/RadixSort.hpp
@@ -29,7 +29,7 @@ public:
         const int radix = 1 << radix_bits;
         const int radix_mask = radix - 1;
         const int num_passes = (num_bits + radix_bits - 1) / radix_bits; // Number of passes needed.
-        const int num_threads = std::thread::hardware_concurrency();
+        const size_t num_threads = std::thread::hardware_concurrency();
         const size_t num_elements = data.size();
 
         Timer timer;
@@ -39,7 +39,7 @@ public:
             timer.start();
         }
 
-        std::vector<std::vector<int>> counts_by_thread(num_threads, std::vector<int>(radix, 0));
+        std::vector<std::vector<size_t>> counts_by_thread(num_threads, std::vector<size_t>(radix, 0));
         std::vector<std::thread> threads;
         const int num_elements_per_thread = static_cast<int>(num_elements / num_threads);
 
@@ -53,10 +53,10 @@ public:
             if (verbose)
                 countPhaseTimer.start("Count phase");
 
-            for (int t = 0; t < num_threads; ++t) {
+            for (size_t t = 0; t < num_threads; ++t) {
                 threads.emplace_back([&, t]() {
                     // Reset counts
-                    for (int r = 0; r < radix; ++r)
+                    for (size_t r = 0; r < radix; ++r)
                         counts_by_thread[t][r] = 0;
                     
                     size_t start = num_elements_per_thread * t;
@@ -77,23 +77,23 @@ public:
             }
 
             // Merge counts to global counts.
-            std::vector<uint32_t> counts(radix, 0);
-            for (int t = 0; t < num_threads; ++t) {
-                for (int r = 0; r < radix; ++r)
+            std::vector<size_t> counts(radix, 0);
+            for (size_t t = 0; t < num_threads; ++t) {
+                for (size_t r = 0; r < radix; ++r)
                     counts[r] += counts_by_thread[t][r];
             }
 
             // Global prefix sum.
-            std::vector<uint32_t> offsets(radix, 0);
-            for (int i = 1; i < radix; ++i)
+            std::vector<size_t> offsets(radix, 0);
+            for (size_t i = 1; i < radix; ++i)
                 offsets[i] = offsets[i - 1] + counts[i - 1];
 
             // Compute per-thread offsets.
-            std::vector<std::vector<int>> offsets_for_thread(num_threads, std::vector<int>(radix, 0));
-            for (int r = 0; r < radix; ++r)
+            std::vector<std::vector<size_t>> offsets_for_thread(num_threads, std::vector<size_t>(radix, 0));
+            for (size_t r = 0; r < radix; ++r)
                 offsets_for_thread[0][r] = offsets[r];
-            for (int t = 1; t < num_threads; ++t) {
-                for (int r = 0; r < radix; ++r)
+            for (size_t t = 1; t < num_threads; ++t) {
+                for (size_t r = 0; r < radix; ++r)
                     offsets_for_thread[t][r] = offsets_for_thread[t - 1][r] + counts_by_thread[t - 1][r];
             }
             
@@ -105,13 +105,13 @@ public:
             Timer redistributionTimer;
             if (verbose)
                 redistributionTimer.start("Redistribution phase");
-            for (int t = 0; t < num_threads; ++t) {
+            for (size_t t = 0; t < num_threads; ++t) {
                 threads.emplace_back([&, t]() {
                     size_t start = num_elements_per_thread * t;
                     size_t end = (t == num_threads - 1) ? num_elements : num_elements_per_thread * (t + 1);
                     for (size_t i = start; i < end; ++i) {
                         KeyType key = (data[i].*key_extractor_ >> shift) & radix_mask;
-                        int outpos = offsets_for_thread[t][key]++;
+                        size_t outpos = offsets_for_thread[t][key]++;
                         buffer[outpos] = data[i];
                     }
                 });
